+++ src/confluent_kafka/admin/__init__.py
@@ -561,7 +561,7 @@ class AdminClient (_AdminClientImpl):
                   Default: `socket.timeout.ms/1000.0`
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
         :param bool validate_only: If true, the request is only validated
                   without creating the topic. Default: False
 
@@ -595,7 +595,7 @@ class AdminClient (_AdminClientImpl):
                   Default: `socket.timeout.ms/1000.0`
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each topic, keyed by the topic name.
                   The future result() method returns None.
@@ -634,7 +634,7 @@ class AdminClient (_AdminClientImpl):
                   Default: `socket.timeout.ms/1000.0`
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
         :param bool validate_only: If true, the request is only validated
                   without creating the partitions. Default: False
 
@@ -668,7 +668,7 @@ class AdminClient (_AdminClientImpl):
         :param list(ConfigResource) resources: Resources to get the configuration for.
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each resource, keyed by the ConfigResource.
                   The type of the value returned by the future result() method is
@@ -712,7 +712,7 @@ class AdminClient (_AdminClientImpl):
         :param list(ConfigResource) resources: Resources to update configuration of.
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`.
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`.
         :param bool validate_only: If true, the request is validated only,
                   without altering the configuration. Default: False
 
@@ -745,7 +745,7 @@ class AdminClient (_AdminClientImpl):
         :param list(ConfigResource) resources: Resources to update configuration of.
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`.
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`.
         :param bool validate_only: If true, the request is validated only,
                   without altering the configuration. Default: False
         :param int broker: Broker id to send the request to. When
@@ -777,7 +777,7 @@ class AdminClient (_AdminClientImpl):
                          to create.
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each ACL binding, keyed by the :class:`AclBinding` object.
                   The future result() method returns None on success.
@@ -816,7 +816,7 @@ class AdminClient (_AdminClientImpl):
                   that is a prefix of the given resource name
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A future returning a list(:class:`AclBinding`) as result
 
@@ -851,7 +851,7 @@ class AdminClient (_AdminClientImpl):
                   that is a prefix of the given resource name
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each ACL binding filter, keyed by the :class:`AclBindingFilter` object.
                   The future result() method returns a list of :class:`AclBinding`.
@@ -878,7 +878,7 @@ class AdminClient (_AdminClientImpl):
 
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
         :param set(ConsumerGroupState) states: only list consumer groups which are currently in
                   these states.
 
@@ -915,7 +915,7 @@ class AdminClient (_AdminClientImpl):
         :param bool include_authorized_operations: If True, fetches group AclOperations. Default: False
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each group, keyed by the group_id.
                   The future result() method returns :class:`ConsumerGroupDescription`.
@@ -948,7 +948,7 @@ class AdminClient (_AdminClientImpl):
         :param bool include_authorized_operations: If True, fetches topic AclOperations. Default: False
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each topic, keyed by the topic.
                   The future result() method returns :class:`TopicDescription`.
@@ -982,7 +982,7 @@ class AdminClient (_AdminClientImpl):
         :param bool include_authorized_operations: If True, fetches topic AclOperations. Default: False
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A future returning description of the cluster as result
 
@@ -1006,7 +1006,7 @@ class AdminClient (_AdminClientImpl):
         :param list(str) group_ids: List of group_ids which need to be deleted.
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each group, keyed by the group_id.
                   The future result() method returns None.
@@ -1043,7 +1043,7 @@ class AdminClient (_AdminClientImpl):
         :param bool require_stable: If True, fetches stable offsets. Default: False
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each group, keyed by the group id.
                   The future result() method returns :class:`ConsumerGroupTopicPartitions`.
@@ -1076,7 +1076,7 @@ class AdminClient (_AdminClientImpl):
                     partition; and corresponding offset to be updated.
         :param float request_timeout: The overall request timeout in seconds,
                   including broker lookup, request transmission, operation time
-                  on broker, and response. Default: `socket.timeout.ms*1000.0`
+                  on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures for each group, keyed by the group id.
                   The future result() method returns :class:`ConsumerGroupTopicPartitions`.
@@ -1126,7 +1126,7 @@ class AdminClient (_AdminClientImpl):
                to describe all user's credentials.
         :param float request_timeout: The overall request timeout in seconds,
                including broker lookup, request transmission, operation time
-               on broker, and response. Default: `socket.timeout.ms*1000.0`
+               on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: In case None is passed it returns a single future.
                   The future yields a dict[str, UserScramCredentialsDescription]
@@ -1162,7 +1162,7 @@ class AdminClient (_AdminClientImpl):
                The pair (user, mechanism) must be unique among alterations.
         :param float request_timeout: The overall request timeout in seconds,
                including broker lookup, request transmission, operation time
-               on broker, and response. Default: `socket.timeout.ms*1000.0`
+               on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures keyed by user name.
                   The future result() method returns None or
@@ -1193,7 +1193,7 @@ class AdminClient (_AdminClientImpl):
                querying.
         :param float request_timeout: The overall request timeout in seconds,
                including broker lookup, request transmission, operation time
-               on broker, and response. Default: `socket.timeout.ms*1000.0`
+               on broker, and response. Default: `socket.timeout.ms/1000.0`
 
         :returns: A dict of futures keyed by TopicPartition.
                   The future result() method returns ListOffsetsResultInfo
@@ -1225,25 +1225,25 @@ class AdminClient (_AdminClientImpl):
     def delete_records(self, topic_partition_offsets, **kwargs):
         """
         Deletes all the records before the specified offsets (not including),
-        in the specified Topic and Partitions.
+        in the specified topics and partitions.
 
-        :param list(TopicPartitions) topic_partition_offsets: A list of
-               TopicPartition objects having `offset` field set to the offset
+        :param list(TopicPartition) topic_partition_offsets: A list of
+               :class:`.TopicPartition` objects having `offset` field set to the offset
                before which all the records should be deleted.
                `offset` can be set to :py:const:`OFFSET_END` (-1) to delete all records
                in the partition.
         :param float request_timeout: The overall request timeout in seconds,
                including broker lookup, request transmission, operation time
-               on broker, and response. Default: `socket.timeout.ms*1000.0`
+               on broker, and response. Default: `socket.timeout.ms/1000.0`
         :param float operation_timeout: The operation timeout in seconds,
-               controlling how long the delete_records request will block
+               controlling how long the `delete_records` request will block
                on the broker waiting for the record deletion to propagate
                in the cluster. A value of 0 returns immediately.
                Default: `socket.timeout.ms/1000.0`
 
-        :returns: A dict of futures keyed by the TopicPartition.
-                  The future result() method returns DeletedRecords
-                  or raises KafkaException
+        :returns: A dict of futures keyed by the :class:`.TopicPartition`.
+                  The future result() method returns :class:`.DeletedRecords`
+                  or raises :class:`.KafkaException`
 
         :rtype: dict[TopicPartition, future]