+++ CHANGELOG.md
@@ -8,7 +8,7 @@ v2.5.0 is a feature release with the following features, fixes and enhancements:
  - Added an example to show the usage of the custom logger with `AdminClient` 
  - Removed usage of `strcpy` to enhance security of the client (#1745)
  - Fixed invalid write in `OAUTHBEARER/OIDC` extensions copy (#1745)
- - Fixed documentation for default value of operation_timeout in create_topics, delete_topics, create_partitions apis (#1710)
+ - Fixed documentation for default value of `operation_timeout` and `request_timeout` in various Admin apis (#1710)
  - Fixed an issue related to import error of `TopicCollection` and `TopicPartitionInfo` classes when importing through other module like mypy.
  - Fixed a segfault when `commit` or `store_offsets` consumer method is called incorrectly with errored Message object
  - Fixed `logger` not working when provided as an argument to `AdminClient` instead of a configuration property